Share models.AddressYears to remove duplication
