+++ app/forms/register/establishers/individual/AddressYearsFormProvider.scala
@@ -17,9 +17,10 @@
 package forms.register.establishers.individual
 
 import javax.inject.Inject
+
 import forms.mappings.Mappings
+import models.AddressYears
 import play.api.data.Form
-import models.register.establishers.individual.AddressYears
 
 class AddressYearsFormProvider @Inject() extends Mappings {
