+++ app/identifiers/register/establishers/company/director/CompanyDirectorAddressYearsId.scala
@@ -18,10 +18,10 @@ package identifiers.register.establishers.company.director
 
 import identifiers._
 import identifiers.register.establishers.EstablishersId
-import models.register.establishers.company.director.CompanyDirectorAddressYears
+import models.AddressYears
 import play.api.libs.json.JsPath
 
-case class CompanyDirectorAddressYearsId(establisherIndex: Int, directorIndex: Int) extends TypedIdentifier[CompanyDirectorAddressYears] {
+case class CompanyDirectorAddressYearsId(establisherIndex: Int, directorIndex: Int) extends TypedIdentifier[AddressYears] {
   override def path: JsPath = EstablishersId.path \ establisherIndex \ "director" \ directorIndex \ CompanyDirectorAddressYearsId.toString
 }
