+++ app/views/register/establishers/company/director/companyDirectorAddressYears.scala.html
@@ -18,7 +18,7 @@
 @import uk.gov.hmrc.play.views.html._
 @import controllers.register.establishers.company.director.routes._
 @import models.Mode
-@import models.register.establishers.company.director.CompanyDirectorAddressYears
+@import models.AddressYears
 
 @(appConfig: FrontendAppConfig, form: Form[_], mode: Mode, establisherIndex: Index, directorIndex: Index)(implicit request: Request[_], messages: Messages)
 
@@ -39,7 +39,7 @@
             field = form("value"),
             legend = messages("messages__company_director_address_years__title"),
             legendClass = Some("visually-hidden"),
-            inputs = CompanyDirectorAddressYears.options
+            inputs = AddressYears.options
         )
 
         @components.submit_button()
