+++ app/views/register/establishers/individual/addressYears.scala.html
@@ -18,7 +18,7 @@
 @import uk.gov.hmrc.play.views.html._
 @import controllers.register.establishers.individual.routes._
 @import models.{Mode}
-@import models.register.establishers.individual.AddressYears
+@import models.AddressYears
 
 @(appConfig: FrontendAppConfig, form: Form[_], mode: Mode, index: Index)(implicit request: Request[_], messages: Messages)
