+++ test/controllers/register/establishers/company/CompanyAddressYearsControllerSpec.scala
@@ -16,20 +16,18 @@
 
 package controllers.register.establishers.company
 
+import connectors.FakeDataCacheConnector
+import controllers.ControllerSpecBase
+import controllers.actions._
+import forms.register.establishers.company.AddressYearsFormProvider
+import identifiers.register.establishers.EstablishersId
+import identifiers.register.establishers.company.CompanyAddressYearsId
+import models.{AddressYears, Index, NormalMode}
 import play.api.data.Form
 import play.api.libs.json.Json
-import uk.gov.hmrc.http.cache.client.CacheMap
-import utils.FakeNavigator
-import connectors.FakeDataCacheConnector
-import controllers.actions._
 import play.api.test.Helpers._
-import forms.register.establishers.company.AddressYearsFormProvider
-import identifiers.register.establishers.company.CompanyAddressYearsId
-import models.{Index, NormalMode}
+import utils.FakeNavigator
 import views.html.register.establishers.company.companyAddressYears
-import controllers.ControllerSpecBase
-import identifiers.register.establishers.EstablishersId
-import models.register.establishers.company.CompanyAddressYears
 
 class CompanyAddressYearsControllerSpec extends ControllerSpecBase {
 
@@ -50,7 +48,7 @@ class CompanyAddressYearsControllerSpec extends ControllerSpecBase {
     EstablishersId.toString -> Json.arr(
       Json.obj(
         CompanyAddressYearsId.toString ->
-          CompanyAddressYears.options.head.value.toString
+          AddressYears.options.head.value.toString
       )
     )
   )
@@ -66,7 +64,7 @@ class CompanyAddressYearsControllerSpec extends ControllerSpecBase {
     "populate the view correctly on a GET when the question has previously been answered" in {
       val getRelevantData = new FakeDataRetrievalAction(Some(validData))
       val result = controller(getRelevantData).onPageLoad(NormalMode, firstIndex)(fakeRequest)
-      contentAsString(result) mustBe viewAsString(form.fill(CompanyAddressYears.values.head))
+      contentAsString(result) mustBe viewAsString(form.fill(AddressYears.values.head))
     }
 
     "redirect to session expired from a GET when the index is invalid" ignore {
@@ -77,7 +75,7 @@ class CompanyAddressYearsControllerSpec extends ControllerSpecBase {
     }
 
     "redirect to the next page when valid data is submitted" in {
-      val postRequest = fakeRequest.withFormUrlEncodedBody(("value", CompanyAddressYears.options.head.value))
+      val postRequest = fakeRequest.withFormUrlEncodedBody(("value", AddressYears.options.head.value))
       val result = controller().onSubmit(NormalMode, firstIndex)(postRequest)
       status(result) mustBe SEE_OTHER
       redirectLocation(result) mustBe Some(onwardRoute.url)
@@ -98,7 +96,7 @@ class CompanyAddressYearsControllerSpec extends ControllerSpecBase {
     }
 
     "redirect to Session Expired for a POST if no existing data is found" in {
-      val postRequest = fakeRequest.withFormUrlEncodedBody(("value", CompanyAddressYears.options.head.value))
+      val postRequest = fakeRequest.withFormUrlEncodedBody(("value", AddressYears.options.head.value))
       val result = controller(dontGetAnyData).onSubmit(NormalMode, firstIndex)(postRequest)
       status(result) mustBe SEE_OTHER
       redirectLocation(result) mustBe Some(controllers.routes.SessionExpiredController.onPageLoad().url)
